dom/tests/mochitest/webcomponents/test_bug1276240.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1276240
-->
<head>
  <title>Test for Bug 1276240</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank"
	href="https://bugzilla.mozilla.org/show_bug.cgi?id=1276240">Mozilla Bug 1276240</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 1276240 **/
// when passing null for the second argument, it would be ignored

function test() {
  var e = document.createElement("p", null);
  is(e.getAttribute("is"), null);

  e = document.createElement("p", undefined);
  is(e.getAttribute("is"), null);

  e = document.createElementNS("http://www.w3.org/1999/xhtml", "p", null);
  is(e.getAttribute("is"), null);

  e = document.createElementNS("http://www.w3.org/1999/xhtml", "p", undefined);
  is(e.getAttribute("is"), null);
}

function runTest() {
  test();
  SimpleTest.finish();
}

// test with webcomponents enabled
test();

// test with webcomponents disabled
SimpleTest.waitForExplicitFinish();
SpecialPowers.pushPrefEnv(
  { 'set': [["dom.webcomponents.enabled", false]]}, runTest);

</script>
</pre>
</body>
</html>
